+++ code_to_pdf/options/__init__.py
@@ -2,7 +2,9 @@
 
 from pydantic import BaseModel
 
-from code_to_pdf.options.custom_types import FontSize, Margin, PygmentsStyle
+from code_to_pdf.options.custom_types import (
+    FontSize, Margin, PaperSize, PygmentsStyle
+)
 
 short_aliases = {"style": "s"}
 
@@ -17,6 +19,7 @@ class Options(BaseModel):
         linenos: Whether or not to display line numbers.
         linenostep: If `linenos` is enabled, print every n-th line number.
         page_numbers: Whether or not to print page numbers.
+        paper_size: Paper size of resulting document.
         top_margin: Document top margin.
         bottom_margin: Document bottom margin.
         left_margin: Document left margin.
@@ -31,6 +34,8 @@ class Options(BaseModel):
     linenostep: int = 0
 
     page_numbers: bool = False
+
+    paper_size: PaperSize = "a4paper"
 
     top_margin: Margin = "0.4in"
     bottom_margin: Margin = "0.4in"